Add proccessing and renaming to user pic
