+++ frontend/src/components/AddItems/AddItemForm.jsx
@@ -16,6 +16,7 @@ import Typography from '@mui/material/Typography';
 import { useEffect, useState } from 'react';
 
 import { IMG_LOGO } from '../../globals/urls';
+import { isImage } from '../../utils';
 import { ImageAvatar } from '../ImageAvatar';
 
 const types = [
@@ -46,22 +47,14 @@ export const AddItemForm = ({
     }
   }, [itemFormData.imgpic]);
 
-  // Checks if the uploaded file is an image based on mimetype
-  // Stricter checking will be done on the server side
-  const isImage = (e) => {
-    const file = e.target.files[0];
-    const fileType = file.type.split('/')[0];
-    if (fileType !== 'image') {
-      alert('Please upload an image file');
-      e.target.value = '';
-      return false;
-    }
-    return true;
-  };
-
   // Handles the file input change
   const onSelectFile = (e) => {
-    if (!e.target.files || e.target.files.length === 0 || !isImage(e)) {
+    if (
+      !e.target.files ||
+      e.target.files.length === 0 ||
+      !isImage(e.target.files[0])
+    ) {
+      e.target.value = '';
       return;
     }
     const imageFile = e.target.files[0];
